refactor: 예약 등록 시 결제가 되는 테스트 케이스 클래스 분리
